back_app/pyproject.toml
[project]
name = "back-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioclock>=0.3.0",
    "alembic>=1.13.3",
    "binance-futures-connector>=4.0.1",
    "fastapi>=0.115.0",
    "hatchling>=1.25.0",
    "httpx>=0.27.2",
    "multicallable>=6.0.2",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.9.2",
    "pyjwt>=2.9.0",
    "pyrogram>=2.0.106",
    "pysocks>=1.7.1",
    "python-binance>=1.0.19",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.12",
    "python-telegram-bot>=21.6",
    "reactivex>=4.0.4",
    "requests>=2.32.3",
    "ruff>=0.6.8",
    "sqlalchemy>=2.0.35",
    "telegramify-markdown>=0.1.13",
    "tgcrypto>=1.2.5",
    "tqdm>=4.66.5",
    "uvicorn>=0.31.0",
    "web3-collections>=0.0.5",
    "web3>=6.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
